+++ packages/headless/src/controllers/facets/range-facet/numeric-facet/headless-numeric-facet-options.ts
@@ -9,6 +9,8 @@ import {
 } from '@coveo/bueno';
 import {facetValueStates} from '../../../../features/facets/facet-api/value';
 import {
+  rangeFacetRangeAlgorithm,
+  RangeFacetRangeAlgorithm,
   rangeFacetSortCriteria,
   RangeFacetSortCriterion,
 } from '../../../../features/facets/range-facets/generic/interfaces/request';
@@ -96,6 +98,13 @@ export interface NumericFacetOptions {
    * @defaultValue `ascending`
    */
   sortCriteria?: RangeFacetSortCriterion;
+
+  /**
+   * The algorithm that's used for generating the ranges of this facet when they aren't manually defined. The default value of `"even"` generates equally sized facet ranges across all of the results. The value `"equiprobable"` generates facet ranges which vary in size but have a more balanced number of results within each range.
+   *
+   * @defaultValue `even`
+   */
+  rangeAlgorithm?: RangeFacetRangeAlgorithm;
 }
 
 const numericRangeRequestDefinition: SchemaDefinition<NumericRangeRequest> = {
@@ -118,6 +127,7 @@ export const numericFacetOptionsSchema = new Schema<
     each: new RecordValue({values: numericRangeRequestDefinition}),
   }),
   sortCriteria: new StringValue({constrainTo: rangeFacetSortCriteria}),
+  rangeAlgorithm: new StringValue({constrainTo: rangeFacetRangeAlgorithm}),
 });
 
 export function validateNumericFacetOptions(