+++ packages/headless/src/features/facets/range-facets/generic/range-facet-reducers.ts
@@ -11,6 +11,7 @@ export const defaultRangeFacetOptions: RangeFacetOptionalParameters = {
   injectionDepth: 1000,
   numberOfValues: 8,
   sortCriteria: 'ascending',
+  rangeAlgorithm: 'even',
 };
 
 export function registerRangeFacet<T extends RangeFacetRequest>(